## Description
Changing the hello world example to adapt to the changes made on the
following roles:

- install_skupper
- deploy_workload

## Motivation and Context

This change is needed to use the new variables defined at the roles that
will allow the installation of skupper to work properly and also
identify either the workload deploy and skupper deploy to be identified
by it's labels and count.

## How Has This Been Tested?

Set the cluster details and the kubeconfig to ~/.kune/config in order to
use the example default.

```bash
make build
make hello
```

## Types of Changes

- [X] Bug fix
- [X] New feature

## Checklist:

- [X] My code follows the style guidelines of this project
- [X] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas

Signed-off-by: Rafael Zago <rafaelvzago@gmail.com>